requirements.txt
pika 
python-dotenv 
redis 
uuid 
numpy
graypy
mtcnn